yfinance-api/requirements.txt
yfinance==0.2.12
mysql-connector-python==8.0.32
schedule==1.1.0